% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accidents_sample}
\alias{accidents_sample}
\alias{tgvejsVersion}
\title{Version of the tgvejs npm package bundled in `tgver`}
\format{
A character vector
}
\description{
Version of the tgvejs npm package bundled in `tgver`
}
\note{
This was generated using the script in the
`data` directory (`tgver.R` file).
}
\examples{
\donttest{
tgver::version
}
}
\keyword{version}
