% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgve.R
\name{background_run}
\alias{background_run}
\title{Internal helper function to run a `plumber` instance on specific
host and port.}
\usage{
background_run(server, port = 8000, host = "127.0.0.1")
}
\description{
Internal helper function to run a `plumber` instance on specific
host and port.
}
