% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{build}
\alias{build}
\title{Set TGVE variables.}
\usage{
build(path, ...)
}
\arguments{
\item{path}{to be passed to `replace` function}

\item{...}{any or all of the TGVE API variables to replace/add values to.}
}
\description{
The placeholders in the bundled version are named with patter: "tgver.*" in
the main component (eatlas-template). They can be replaced using the
`replace()` function.
}
