% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_tgve}
\alias{copy_tgve}
\title{Internal helper function to:
1. copy the bundled zip
2. unzip
3. cleanup as required.
TODO: return a value?}
\usage{
copy_tgve(path, over.write = TRUE)
}
\arguments{
\item{path}{character path of TGVE instance}

\item{over.write}{boolean whether to cleanup the instance in `path`}
}
\description{
Internal helper function to:
1. copy the bundled zip
2. unzip
3. cleanup as required.
TODO: return a value?
}
