% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\title{Explore an sf R object using Turing eAtlas.}
\usage{
explore(sf = NULL)
}
\arguments{
\item{sf}{a valid sf object that can be converted to geojson}
}
\description{
Explore an sf R object using Turing eAtlas.
}
\examples{
\dontrun{
explore()
}
}
