% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_url}
\alias{get_url}
\title{Helper function to generate URLs}
\usage{
get_url(base = "http://127.0.0.1:8000", ...)
}
\arguments{
\item{base}{character URL defaults to `http://127.0.0.1:8000`}

\item{...}{any or all of the TGVE API variables to replace/add values to.}
}
\value{
character URL generated from `base` and `tgver::`
}
\description{
The function can generate a URL based on a base URL and as many as TGVE API
variables provided to the function.
}
\examples{
{
url = get_url(dark="false")
url == "http://127.0.0.1:8000?dark=false"
url
url = get_url()
url
}

}
