% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{help}
\alias{help}
\title{Function to explore available API variables, their types and examples of
using them in the TGVE}
\usage{
help()
}
\value{
no object is returned
}
\description{
Function to explore available API variables, their types and examples of
using them in the TGVE
}
\examples{
{
help()
}
}
