% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_valid_url}
\alias{is_valid_url}
\title{Good enough regex to sanitize URLs}
\usage{
is_valid_url(string)
}
\arguments{
\item{string}{must be valid vector of URLs}
}
\description{
The task of checking a URL is "hard", see this by J. Hester:
https://cran.r-project.org/web/packages/rex/vignettes/url_parsing.html
To avoid having a dependency for now, let us not use "rex"  R package.
To try and understand the regex please see this gist which includes
a breakdown of the regex:
https://gist.github.com/dperini/729294
}
