% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{list_api_files}
\alias{list_api_files}
\title{Function to find what files may contain TGVE
API variables for functions like `file_replace`
to consume.}
\usage{
list_api_files(path = NULL)
}
\arguments{
\item{path}{where TGVE instance is located.}
}
\description{
Function to find what files may contain TGVE
API variables for functions like `file_replace`
to consume.
}
