% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{openURL}
\alias{openURL}
\title{Internal helper function to "browse" a URL.}
\usage{
openURL(
  url = NULL,
  host = "127.0.0.1",
  port = 8000,
  browser = FALSE,
  path = "",
  protocol = "http://"
)
}
\arguments{
\item{url}{character url, if given and valid other
parameters will be ignored.}

\item{host}{character host to pass to plumber}

\item{port}{integer port to pass to plumber}

\item{browser}{Boolean whether to specifically launch a browser}

\item{path}{character path to TGVE instance}

\item{protocol}{character protocol, this may change}
}
\description{
Internal helper function to "browse" a URL.
}
