% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{replace}
\alias{replace}
\title{Helper function to replace particular variable in bundled minified TGVE JS.
The difference between this function and base::gsub is only finding those
files that contain the pattern, use gsub on them and write the fails back to
where they are.}
\usage{
replace(path = NULL, pattern, replacement)
}
\arguments{
\item{path}{where instance of TGVE is located.}

\item{pattern}{the pattern to replace using gsub.}

\item{replacement}{the string to replace pattern with using gsub.}
}
\description{
Helper function to replace particular variable in bundled minified TGVE JS.
The difference between this function and base::gsub is only finding those
files that contain the pattern, use gsub on them and write the fails back to
where they are.
}
