% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{serve}
\alias{serve}
\title{Launch a webserver via plumber to serve TGVE static files.
Currently the function also copies the bundled
TGVE version into a temporary directory.}
\usage{
serve(
  port = 8000,
  dir = Sys.getenv("TEMP_DIR_ENV"),
  host = "127.0.0.1",
  background = FALSE
)
}
\arguments{
\item{port}{to serve from}

\item{host}{host to pass to plumber default `http://127.0.0.1`
to `FALSE`.}

\item{background}{run the R process in the background using callr,
defaults to `TRUE`.}

\item{file}{location of plumber.R file used by plumber}
}
\description{
Launch a webserver via plumber to serve TGVE static files.
Currently the function also copies the bundled
TGVE version into a temporary directory.
}
\examples{
{
}
}
