% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Helper function to setup a local directory instead of `tempInstance`.}
\usage{
setup(path = NULL, create = TRUE)
}
\arguments{
\item{path}{Character URI to copy tgve instance in.}

\item{create}{Boolean to create new directory at path, defaults to `TRUE`.}
}
\description{
Helper function to setup a local directory instead of `tempInstance`.
}
\examples{
{
p = file.path(tempdir(), "tgve")
setup(p)
}
}
