% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgve.R
\name{tgve}
\alias{tgve}
\title{Launch a webserver via plumber to serve TGVE static files.}
\usage{
tgve(
  dir = Sys.getenv("TEMP_DIR_ENV"),
  port = 8000,
  host = "127.0.0.1",
  background = FALSE
)
}
\arguments{
\item{dir}{location of TGVE dir to be served by plumber.}

\item{port}{to serve from.}

\item{host}{host to pass to plumber default `http://127.0.0.1`
to `FALSE`.}

\item{background}{run the R process in the background using callr,
defaults to `TRUE`.}
}
\description{
The function accepts a `dir` to get a directory containing an instance of
TGVE, by default this is done via `TEMP_DIR_ENV` env variable. If neither
is given then the function copies a clean copy of the bundled
TGVE version into a temporary directory.
}
\examples{
{
}
}
