% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/utils.R
\docType{data}
\name{version}
\alias{version}
\alias{tgvejsVersion}
\title{Version of the tgvejs npm package bundled in `tgver`}
\format{
A character vector
}
\description{
Version of the tgvejs npm package bundled in `tgver`

Package version included as data
}
\note{
This was generated using the script in the
`data` directory (`tgver.R` file).
}
\examples{
{
tgver::version
}
}
\author{
L Hama \email{l.hama@leeds.ac.uk}
}
\keyword{data}
\keyword{version}
