% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Set the TGVE variables.}
\usage{
build(path, ...)
}
\arguments{
\item{path}{to be passed to `file_replace` function}

\item{...}{any or all of the TGVE API variables to replace/add values to.}
}
\description{
The placeholders in the bundled version are named with pattern: "tgver.*" in
the main component (eatlas-template). They can be replaced using the
`file_replace()` function.
}
\examples{
\dontrun{
build(file.path(tempdir(), "tgve"))
}
}
