% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_dir.R
\name{explore_dir}
\alias{explore_dir}
\title{A function to read and serve at least one file from a directory.}
\usage{
explore_dir(path, background = FALSE)
}
\arguments{
\item{path}{character of an existing directory with at least one file in}

\item{background}{logical value to run instance in `callr`}
}
\description{
Current version only tries to:
1. Find two files, one geography.geojson and other data.csv and
pass them to TGVE as `defaultURL` and `geographyURL` with either
matching columns.
2. TODO: Find only one file, pass this to `explore_file`.
}
