% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_sf.R
\name{explore_sf}
\alias{explore_sf}
\title{Explore an sf R object using Turing eAtlas.}
\usage{
explore_sf(sf = NULL, background = FALSE)
}
\arguments{
\item{sf}{a valid sf object that can be converted to geojson}

\item{background}{Boolean to decide whether plumber
should run in the background}
}
\description{
Explore an sf R object using Turing eAtlas.
}
\examples{
\dontrun{
explore_sf()
}
}
