% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_sf.R
\name{explore_sf}
\alias{explore_sf}
\title{Explore an sf R object using TGVE.}
\usage{
explore_sf(
  sf = NULL,
  background = FALSE,
  static = FALSE,
  path = tempInstance()
)
}
\arguments{
\item{sf}{a valid sf object that can be converted to geojson}

\item{background}{boolean to decide whether plumber should run in the
background}

\item{static}{boolean to decide whether data is written to disk and self
contained application is built}

\item{path}{path of a TGVE instance, defaults to one in `tempdir()`}
}
\description{
Explore an sf R object using TGVE.
}
\examples{
\dontrun{
explore_sf()
}
}
