% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{openURL}
\alias{openURL}
\title{Internal helper function to "browse" a URL.}
\usage{
openURL(
  host = "127.0.0.1",
  port = 8000,
  browser = FALSE,
  path = "",
  protocol = "http://"
)
}
\description{
Internal helper function to "browse" a URL.
}
