% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Setup an instance.}
\usage{
setup(path = NULL, create = TRUE)
}
\arguments{
\item{path}{Character URI to copy tgve instance in.}

\item{create}{Boolean to create new directory at path, defaults to `TRUE`.}
}
\value{
no value returned
}
\description{
This function initializes an instance of TGVE for permanent use compared with
what \link[tgver]{tgve_server} which relies on a `tempdir` based instance.
It requires a path.
}
\examples{
{
p = file.path(tempdir(), "tgve")
setup(p)
}
}
