% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgve.R
\name{tgve}
\alias{tgve}
\title{Open static TGVE instance}
\usage{
tgve(path = Sys.getenv("TEMP_path_ENV"), browse = TRUE)
}
\arguments{
\item{path}{character directory of a current instance to browse. Defaults to
`TEMP_PATH_ENV` environment variable.}

\item{browse}{boolean to decide whether to browse the instance or not.}
}
\value{
directory of the new instance if `path` is not provided.
}
\description{
This is the main and most basic function to run an instance of TGVE without
back-end. Compared to \link[tgver]{tgve_server}, this function only uses
\link[tgver]{build} to setup an instance then opens the entry HTML file.
If a path of an instance is provided it opens it, otherwise creates an
instance at the `tempdir`.
}
\examples{
{
tgve()
# just get the path of the HTML
p = tgve(browse = FALSE)
file.exists(p)
}

}
